functions/version3/biginference/requirements.txt
opencv-python-headless
numpy
minio==7.1.1
Pillow==8.4.0
tensorflow